Complementary fixes for files previous edited for MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31639 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/alg/gdalrasterize.cpp
@@ -380,7 +380,7 @@ gv_rasterize_one_shape( unsigned char *pabyChunkBuf, int nYOff,
       case wkbPoint:
       case wkbMultiPoint:
         GDALdllImagePoint( sInfo.nXSize, nYSize, 
-                           aPartSize.size(), &(aPartSize[0]), 
+                           static_cast<int>(aPartSize.size()), &(aPartSize[0]), 
                            &(aPointX[0]), &(aPointY[0]), 
                            (eBurnValueSrc == GBV_UserBurnValue)?
                            NULL : &(aPointVariant[0]),
